+++ client/src/app/cabinet/roles/role-detail/role-detail.component.ts
@@ -2,6 +2,10 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Role } from '../../../models/cabinet/users/role';
 import { RolesService } from '../../../services/cabinet/roles/roles.service';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { WarningConfirmationComponent } from '../../shared/warning-confirmation/warning-confirmation.component';
+import { RedirectService } from '../../../services/cabinet/shared/redirect/redirect.service';
 
 @Component({
   selector: 'app-role-detail',
@@ -10,20 +14,40 @@ import { RolesService } from '../../../services/cabinet/roles/roles.service';
 })
 export class RoleDetailComponent implements OnInit {
   public role: Role;
+  private message: string  = 'Are you sure you want to delete this role?';
 
-  constructor(private rolesService: RolesService, private route: ActivatedRoute, private router: Router) { }
+  constructor(
+    private rolesService: RolesService,
+    private route: ActivatedRoute,
+    private router: Router,
+    private dialog: MatDialog,
+    private snackbar: MatSnackBar,
+    private redirectService: RedirectService
+  ) { }
 
   ngOnInit(): void {
     const id = this.route.snapshot.params['id'];
     this.rolesService.getRoleById(id).subscribe((response) => {
       this.role = response.role;
-      console.log(this.role);
     });
   }
 
   public removeRole(id: number): void {
-    this.rolesService.removeRole(id).subscribe(() => {
-      this.router.navigate(['/cabinet/roles']).then();
+    const dialogRef = this.dialog.open(WarningConfirmationComponent, {
+      width: '400px',
+      height: '200px',
+      data: { message: this.message }
+    });
+    dialogRef.afterClosed().subscribe((dialogResult) => {
+      if (dialogResult) {
+        this.rolesService.removeRole(id).subscribe(() => {
+          this.snackbar.open('Role was deleted!', 'Close', {
+            duration: 2000,
+            verticalPosition: 'top'
+          });
+          this.redirectService.redirect('/cabinet/roles', 2000);
+        });
+      }
     });
   }
